requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
discord.py==0.16.12
gunicorn==19.7.1
heroku==0.1.4
idna==2.6
inflect==0.2.5
lxml==4.1.1
multidict==4.1.0
overwatch-stats==1.1.0
python-dateutil==1.5
requests==2.18.4
urllib3==1.22
websockets==3.4
